+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project are documented in this file.
 - Add endpoint to get city by ID.
 - Add endpoint to get district by ID.
 - Add endpoint to get neighborhood by ID.
+- Add GraphQL for the city query.
 
 ### Refactored